patch out the cairo/gi python module detection.
sadly, gpsd uses scons as a build system, which is python2 only, thus
tries to load py-cairo and py-gobject3 to detect their presence. We
enfore the RUN_DEPENDS on the python3 version of those, so no need for
this check.

to be revisited if scons switches to python3 someday.

Fixes packaging problem when py-cairo/py-gobject3 are missing as
reported by ajacoutot@