Import devel/p5-MooseX-Object-Pluggable
This module is meant to be loaded as a role from Moose-based classes
it will add five methods and four attributes to assist you with the
loading and handling of plugins and extensions for plugins. I
understand that this may pollute your namespace, however I took great
care in using the least ambiguous names possible.

From wen heping <wenheping2000 () hotmail ! com>
OK cwen@