use cython instead of pyrex in mysticmine
pyrex is used for a single .pyx file. Since development on this game seems
to have stopped we could pre-generate a static copy of ai.c so we can drop
the pyrex dependency. However, it turns out that the pyrex code is similar
enough to modern cython, that we can just swap pyrex for cython. This seems
to work...

While here:
- remove pyrex as an RDEP, not sure why that was ever needed
- fix the locale-related crasher so the game doesn't just bail at startup
- update to what looks like the final github commit (from July 2016)
- s/koonsolo/dewitters/ as requested by kmos@
- regenerate PLIST

Tested by myself on amd64

ok kmos@