Mark BROKEN for now.
As seen on http://www.radio-browser.info/gui/#!/ the radio-browser
database API migrated to a new incompatible version and gradio is
abandoned upstream/author moved to a new rust-based project named
shortwave (sigh).

Will eventually dive into vala to adapt to the new API, otherwise this
port is a candidate for the Attic.