otter-browser: move to QtWebEngine
I asked for QtWebKit in Otter to be kept when QtWebEngine was introduced,
because it was the only fully featured browser that really worked on macppc.

Since the last QtWebKit update, it was crashing when running javascript. And
now the rust-powered spidermonkey78 is in its chain of dependencies, so it's
not built anymore on macppc anyway.

OK rsadoswki@, maintainer timeout