webkitgtk4: powerpc-only fixes
1) Fix javascript crashes at runtime
Take a patch from Daniel Kolesa @ Void-PPC to avoid segfaults in consumers as
soon as javascript is used. Keep it powerpc-only for now, since it has not
yet been reviewed by upstream and can have consequences on other 32-bit archs.

2) Fix colors
Colors were off as soon as javascript is involved. Disabling context GL fixes
the issue (but hurts performances). Another patch was needed because of a
missing include when this option is disabled.

Thanks to sthen and aja for hints!
OK aja@ (maintainer)