py-sphinx no longer depends on py-sphinx_rtd_theme
Release 1.4 of py-sphinx made this dependency optional. So adapt the port
to match. Instead we add a BUILD_DEP on py-sphinx_rtd_theme to the 3 ports
that actually depend on it. This may make updating py-sphinx_rtd_theme a
bit easier to update in the future as fewer consumers would need testing.

ok kmos@, sthen@