Import recoll-1.27.12
Recoll is a full text search application, which means that it finds your data
by content rather than by external attributes (like the file name). You specify
words (terms) which should or should not appear in the text you are looking for,
and receive in return a list of matching documents, ordered so that the most
relevant documents will appear first.
<...>

ok sthen@