update to asterisk-18.1.1
AST-2020-003: Remote crash in res_pjsip_diversion -
A crash can occur in Asterisk when a SIP message is received that has a
History-Info header, which contains a tel-uri.

AST-2020-004: Remote crash in res_pjsip_diversion -
A crash can occur in Asterisk when a SIP 181 response is received that
has a Diversion header, which contains a tel-uri.