import libmodsecurity 3.0.4 from Matthias Pitzl; OK benoit@
Comment:
intrusion detection and prevention engine

Description:
ModSecurity is an embeddable web application firewall. It provides
protection from a range of attacks against web applications and allows
for HTTP traffic monitoring and real-time analysis with no changes to
existing infrastructure.

It is also an open source project that aims to make the web application
firewall technology available to everyone.