security/openssl/1.0.2: fix alignment issue on sparc64
As diagnosed by kettenis, running the regress/lib/libssl/interop/openssl
test results in a SIGILL or SIGBUS due to an alignment issue. The reason
for this is that the configure magic fails to pick up -m64 which is
needed for the perlasm to generate the correct flavor of assembly.

None of the approaches for setting variables in the main port Makefile
worked.  Since we already patch out -O3 from Configure, I added -m64
there. The resulting binary seems to work well. The interop tests on
sparc64 pass with this patch.

Many thanks to kettenis for figuring this out and to sthen who helped
me save a lot of time with FLAVOR=no_man.

ok sthen