update to libuv-1.40.0
switch to the py3 version of sphinx; if i'm not mistaken this avoids
the need to build both python 2 and 3 before building cmake. (this is on
the path to building about a third of the ports tree).

tests done on aarch64 i386 amd64, and some deps, ports-gcc, USE_LLD=no
tested on amd64.