Remove qtstyleplugins-5.0.0.20170311
"it was originally ported to 'fix' the qt5 bug with file dialogs when
running within gtk desktop environments, the default gtk3 file dialog
triggers a cpu 100% spin." -- landry

This issue still exists so DO NOT USE GTK3-style (which should not be the
default) for (py)-Qt applications.

Qt5 does not ship a qtconfig utility to configure fonts, icons or styles. You
can use the qt5ct package to do this job.  Don't forget to set the environment
variable QT_QPA_PLATFORMTHEME=qt5ct so that the settings are picked up by Qt
applications

OK kmos, kn and landry