import ports/graphics/openvdb, from deserter666 at danwin1210.me, ok rsadowski
OpenVDB is an Academy Award-winning open-source C++ library comprising a novel
hierarchical data structure and a suite of tools for the efficient storage and
manipulation of sparse volumetric data discretized on three-dimensional grids.
It was developed by DreamWorks Animation for use in volumetric applications
typically encountered in feature film production and is now maintained by the
Academy Software Foundation (ASWF).