Remove security/hitch
Over half a year ago I dropped MAINTAINER on this port due to not using it
any longer.  At that time it was already outdated.  Noone spoke up to
update or even maintain it.

The new (unported) version 1.6.0 already suffers from TLS related build
failures, now there's another problem: it does not build with "-fcommon"
which will become a default compiler option in the tree.

If someone wants to fix both and get an up-to-date version running they
recover it from the attic.

OK tb