Remove biology/nutdb
It fails to build with "-fno-common" and HOMEPAGE indicates that it's
no longer being worked on.

The website points at https://cronometer.com/ (website / mobile app) to be
used in the interim until upstream's NUT rewrite is done.

There is NUTsqlite on the sourceforge site's files section, not sure
but perhaps that is the rewrite talked about on the website. It is
written in Tcl/Tk. (thanks sthen)

OK sthen