Import ruby-argon2 2.0.3
This Ruby Gem provides FFI bindings, and a simplified interface, to the
Argon2 algorithm. Argon2 is the official winner of the Password Hashing
Competition, a several year project to identify a successor to
bcrypt/PBKDF/scrypt methods of securely storing passwords. This is an
independant project and not official from the PHC team.

OK sthen@