Here's a new port for lwjgl3. It's a new port since there are existing things that use the lwjgl 2.x version we already have.
Upstream is not interested in supporting BSD and has expressed hope that
someone would fork the code for BSD support (so they don't have to deal
with it). So this is based upon my fork with OpenBSD support added.
(https://github.com/kmosiejczuk/lwjgl3)

I added most of the code to support OpenBSD with some help from pamela
and phessler. octeep (https://github.com/octeep/lwjgl3) figured out the
last couple of bits that eluded me for so long. Thank you!

The build will try and download needed things to build, including a specific
version of the kotlin compiler and some java libraries from maven. Thus the
tricks of adding them in. I just bundled the maven libraries and put the
distfile on a server of mine.

phessler and pamela helped some with the port (blame the sins on me though)
and both helped test.

input from pamela@ phessler@ thfr@ sthen@

ok phessler@