whack py-sip and friends into shape. the previous ports update ended up with two different versions of py-sip; the old one configured to only produce the module for py-qt5 (in devel/py-sip) and the actual sip in x11/py-sip-qt5, with broken update chain and which didn't entirely make sense.
in particular: py-sip is now actually (a newer version of) sip again,
with py2 support dropped (current versions don't have it). other ports
using this are changed to the newer build system using "sip-build"
where possible.

for the future we will need to move to sip 6.x which drops the older
build system using the "sip5" or "sip" script; other ports (veusz, and
i think qgis) don't work with that yet.