audio/audacous: Add a qt5 FLAVOR.
This enables the new qt5 GUI in a FLAVOR. It's a bit more streamlined,
and it's easier to dock and arrange all of your various audacious
windows.

Although upstream is deprecating the GTK GUI, we decided that we are not
quite ready to fully ditch in favour of only qt5.

Also add a meta package FLAVOR.

This is a team effort between myself and MAINTAINER, bket@.

OK bket@, many thanks!