Unhook mail/avenger
We package 0.8.3 05-Apr-2012 but 0.8.4 in 13-Jul-2013 and 0.8.5 in
08-Oct-2018 were released;  the release tarball's NEWS file states
"changes to make it compile with newer g++ and libcrypto" and
"Manually parse /etc/resolv.conf, rather than attempting to make use of the
res data structure (which at least on OpenBSD no longer actually contains
the addresses of resolvers)".

This port fails to build with "-fno-common" and noone has spoken up on
ports@ to fix and update it.

PS: http://www.mailavenger.org/NEWS is sitll missing 0.8.5 changes, so
best look at the actual tarball.

Unhook now and give it a few days grace period before actual removal.
OK gnezdo for sending it to the attic