import textproc/p5-Pod-Parser, ok afresh1@ (plus I added his patches from the version in base). (also uses the shiny new FIX_CRLF_FILES). This is currently in perl in base but will be removed in 5.32.
----
Pod::Parser is a base class for creating POD filters and translators.  It
handles most of the effort involved with parsing the POD sections from an
input stream, leaving subclasses free to be concerned only with performing
the actual translation of text.

NOTE: This module is considered legacy. Pod-Simple is the usual module for
all things POD.
----