remove x11/golem
Our port of this old window manager is at version 0.0.5, which is
from 2002 (19 years old).  At startup, there is a spew of errors
about undefined symbols and the plugins can't be loaded.  Clearly,
nobody has been using this.

There is a newer development version 0.0.6 from 2007, but it relies
on Linux-only code.

ok kn@ ajacoutot@