update to Dovecot 2.3.14, ok Brad
Note that some components have been removed in this release, summarised
below but check changelogs before updating. https://dovecot.org/doc/NEWS

* Remove autocreate, expire, snarf and mail-filter plugins.

expire and autocreate can be replaced by "autoexpunge" and "auto"
mailbox settings instead. mail-filter was a bit problematic anyway.
snarf is not often used.

* Remove cydir storage driver.

This was meant for benchmarking and stress testing index handling
and is not normally used in a real server.

* Remove XZ/LZMA write support. Read support will be removed in future release.

If you're using XZ/LZMA then disable it for writes in your configuration
(mailboxes can have a mixture of different types of compression, or some
and no compression) and plan to migrate your existing storage to either
no compression or alternative compression.