libexecinfo (libbacktrace): remove -Werror, to allow building with llvm 11 which otherwise fails with newly added warnings:
calling '__builtin_frame_address' with a nonzero argument is unsafe [-Werror,-Wframe-address]
calling '__builtin_return_address' with a nonzero argument is unsafe [-Werror,-Wframe-address]

this doesn't make things any worse, and it seems to function as well as
previously with llvm 11. (there are other discussions about this port,
but they will take longer).