Import x11/xmenu-4.5.4
XMenu is a menu utility for X. XMenu receives a menu specification in stdin,
shows a menu for the user to select one of the options, and outputs the option
selected to stdout. XMenu can be controlled both via mouse and via keyboard.

ok solene@

Port submitted by Guy Godfroy, maintainer.