Import vollkorn-4.105.
Vollkorn is a quiet, modest and well working body copy typeface for bread
and butter use. It has dark and meaty serifs and a bouncing and healthy look
and might be used as body type as well as for headlines or titles. More than
2000 glyphs per font support a wide range of languages in Latin, Cyrillic
and Greek scripts.

"Vollkorn" [pronounced "follkorn"] is German for "wholemeal". It refers to
the old term "Brotschrift" [literally "bread type"] which described the
small fonts for every day use in the days of hand-compositing.

Vollkorn came into being as Friedrich Althausen's first type designing
attempt during his studies at Bauhaus University Weimar, Germany. The
Regular style was initially published in 2005 under a Creative Commons
license. When GoogleFonts launched in 2010 Vollkorn was one the first
twenty featured fonts.

ok sthen@