Import vger-1.06
Vger is a gemini server supporting chroot, virtualhosts,
CGI, default language choice, redirections and MIME types
detection.

Vger design is relying on inetd and a daemon to take care of
TLS.  The idea is to delegate TLS and network to daemons
which proved doing it correctly, so vger takes its request
from stdin and output the result to stdout.

ok thfr@