qutebrowser: fix QtWebEngine version detection, from maintainer following a report from Anindya Mukherjee
qutebrowser needs to identify the version of chromium used in the QtWebEngine
build; this appears to not be easily available via QtWebEngine/PyQtWebEngine
API so it does this by attempting to parse the ELF .so file, however it does
not know about OpenBSD ports shared library versioning.

The fix used here is to search for the highest numbered .so file and use that.