Avoid a crash in libwebrtc threading code
The code on libwebrtc depends on unportable behavior, namely that
pthread_create(3) updates *thread before the child is run.  This happens
on Linux/glibc, FreeBSD and Illumos, but not on OpenBSD or NetBSD.
POSIX only says that *thread is updated after a successful call.

This fixes crashes with Firefox when using Jitsi with several users in
the room.  Input from jmatthew@ and kettenis@, tests by mpi@, ok landry@
(maintainer).