devel/cargo module: add option for libsqlite3-sys crate
libsqlite3-sys only provides "old" bindings for sqlite3 library
(better is 3.7.16). when a crate requires a binding for a recent
version of sqlite3, it is required to use the bundled version to have
proper binding.

add an option MODCARGO_CRATES_SQLITE3_BUNDLED to explictly request it
(and not use the ports version of sqlite3 but build the bundled
version).