Import sysutils/p5-Data-Float 0.013
Description:
This module is about the native floating point numerical data type. A floating
point number is one of the types of datum that can appear in the numeric part of
a Perl scalar. This module supplies constants describing the native floating
point type, classification functions, and functions to manipulate floating point
values at a low level.

Suggestions and ok sthen@