update to MariaDB 10.5.10, from Brad; includes fixes for CVE-2021-2154, CVE-2021-2166 (DoS).
Various hash_XXX functions have been renamed to ma_hashtbl_XXX to fix
an issue with static linking to newer GnuTLS releases;
https://jira.mariadb.org/browse/CONC-543
https://github.com/mariadb-corporation/mariadb-connector-c/commit/d19c7c69269fdf4e2af8943dd86c12e4e1664afd

These functions don't appear to be in the exported API so I have skipped
bumping the library version so that the same update can be used in stable.
I don't expect problems but if we run into any it can be bumped later.