import p5-Test-Lib 0.002 OK sthen@
Comment:
use libraries from a t/lib directory

Description:
Searches upward from the calling module for a directory t with a lib
directory inside it, and adds it to the module search path. Looks upward
up to 5 directories. This is intended to be used in test modules either
directly in t or in a subdirectory to find their included testing
libraries located in t/lib.