Update p5-Moose to 2.2015
Instead of fixing missing dependencies on the old version, used
portgen to update to the new version.  Cleaned up the Makefile
while here.

There is one test that wants a newer version of List::Util than we
have in base, so patching that test out until we update.

Prompted and "sure" from naddy@