Import www/nextcloud_notify_push 0.1.7.
This nextcloud service attempts to solve the issue where Nextcloud clients have
to periodically check the server if any files have been changed. In order to
keep sync snappy, clients wants to check for updates often, which increases the
load on the server.

With many clients all checking for updates a large portion of the server load
can consist of just these update checks.

By providing a way for the server to send update notifications to the clients,
the need for the clients to make these checks can be greatly reduced.

Been using it for some weeks on 6.9 without issues, support for push
notifications landed in nextcloudclient 3.2, and is also present in
recent android/ios clients.

Note that the _notify_push user needs to be in www group to be able to
read /var/www/nextcloud/config.php. The service also requires nextcloud
to be configured against redis, cf pkg/README.

ok gonzalo@, feedback rsadowski@