Update astro/sunclock to 3.57
This is update for sunclock from 3.56 to 3.57.

remove REVISION for the first revision of 3.57.
remove HOMEPAGE since it is unreachable now.
change MASTER_SITES to github with using commit hash.
remove MASTER_SITES0 and :0 from the end of the SUPDISTFILES entries.
remove Xext, Xau and Xdmcp from WANTLIB-main.
replace Sunclock.se with Sunclock.sv in PLIST-main.

advise and ok from sthen@ and jca@