Simplify: stop building the C extension
This C extension is only useful as a fallback on Linux, for direct
syscalls access.  Instead, tweak setup.py so that ctypes properly finds
the inotify* symbols from devel/libinotify instead of libc.

As a result of not building the C extension, this fixes the build on riscv64.

ok jasper@ (maintainer)