Set Puppet-related ports to use Ruby 2.7
Puppet 5 and 6 currently use Ruby 2.7, as that is the default Ruby
version.  Puppet 6 can probably run on Ruby 3.0 after a distpatch,
but Puppet 5 would require additional backporting (how much is
unknown).  Both Puppet 5 and 6 depend on facter and mcollective,
and all must use the same Ruby version. So setting all four ports
to use Ruby 2.7 seems like the simplest way to continue to keep things
working when the default Ruby version is switched to Ruby 3.0.

While here, add install patches for Puppet 5, 6, and mcollective
to build with Ruby 3.0, and add backport a distpatch from Puppet 7
to Puppet 6 to allow Puppet 6 to run on Ruby 3.0.  This will make a
future switch to Ruby 3.0 easier.

These changes should not have a runtime effect on Puppet 5 or 6.
The Puppet 6 port is bumped, due to the backported patch, but that
patch should not change runtime behavior on Ruby 2.7.

review from sebastia@ (Puppet 5 maintainer)
OK kn@ (Puppet 6 maintainer)