import ports/fonts/juliamono, from Laurie Tratt, ok bentley@
JuliaMono is a monospaced typeface designed for programming and in
other text editing environments that require a wide range of
specialist and technical Unicode characters. It comes in 6 weights
(from Light to Black): each has an italic variant; and each occupies
the same amount of horizontal space, so they can be mixed without
losing alignment.