Rollback to boost 1.67.0
There were some errors reported with it. Especially with
net/libtorrent-rasterbar. This meant that net/qbittorrent and net/deluge no
longer wokred. tb@ und mariani@ have found out that it is definitely the boost
update to 1.77. Brad has found out tha the problem can be solved by building
libtorrent-rasterbar with -O0.

Unless we understand the mechanism for the failure and are sure it's a
libtorrent-rasterbar problem rather than a boost one, sthen@ suggested to back
it out.

OK mariani@, sthen@ and Brad