pulseaudio: fix build on !base-clang architectures.
pulseaudio's meson.build effectively uses __builtin_dgettext() to
decide whether linking with -lintl is needed or not: base-clang
doesn't have it, so -lintl is given, while gcc 8.4. provides it,
so meson attempts to build without -lintl, which breaks the build
of pacat and a few other programs.

ok ajacoutot sthen