Import py-puremagic
DESCR:
Puremagic is a pure python module that will identify a file based off
it's magic numbers. It is designed to be minimalistic and inherently
cross platform compatible. It is also designed to be a stand in for
python-magic.

It does NOT try to match files off non-magic string. In other words it
will not search for a string within a certain window of bytes like
others might.


This is a new dependency for news/sabnzbd.

OK sthen@