Import py-rebulk
DESCR:
ReBulk is a python library that performs advanced searches in strings
that would be hard to implement using re module or String methods only.

It includes some features like Patterns, Match, Rule that allows
developers to build a custom and complex string matcher using a readable
and extendable API.


This port is part of a new set of dependencies for sabnzbd.

OK sthen@