move Boost::stacktrace libraries from md to the main package, from Brad.
"Builds on all of the Clang archs and sparc64. It was disabled initially for
a build issue on arm, but that was before switching to Clang.

NetBSD disables context, coroutine and fiber on a few archs like we do, but
not stacktrace and as far as I can see it does not have any MD dependencies."