Import x11/alacritty.
Alacritty is a modern terminal emulator that comes with sensible
defaults, but allows for extensive configuration.

The software is considered to be at a beta level of readiness; there
are a few missing features and bugs to be fixed, but it is already
used by many as a daily driver.

From Eric Auge, thanks! Feedback from sthen@ and tb@.

OK tb@, sthen@