Import games/amnesia-tdd, a first person survival horror game.
Credits:
brynet@ for intial port and upstreaming OpenBSD patches.
shamazmazum for providing an upstream for this port and getting it to
run on FreeBSD.
ok brynet@ thfr@

Amnesia: The Dark Descent, a first person survival horror. A game about
immersion, discovery and living through a nightmare. An experience that
will chill you to the core.