Zap lang/python module usage
Disabling the build/run dependency (--with-python=no) also disables it for
tests, i.e. the configure check is saved and the test suite does not
recheck for python.

Since we won't add a dependency on python, the module is of no use any
longer and we'll let the tests needing it just skip as before.